Move test file to mirror package structure (tests/core/lib/test_ssl.py)
